framework/price_feeds/requirements.txt
websockets>=12.0
aiohttp>=3.9
redis>=5.0
python-dotenv>=1.0
